Version 0.4 is now available!
The main changes include: renaming fields, variables, methods, and other elements for better clarity. The extraction of duplicate pieces of code into separate methods to improve code organization. Additionally, a bug fix has been implemented to prevent the creation of a separate data file with a seed and map size of 0.